update: 修改结构，去掉无用代码

+++ server/socket.js
@@ -24,7 +24,6 @@ class Socket {
         this.ns = instance.of(`/${id}`)
         this.init()
         this.videos = []
-        sockets[id] = this
         this.sequence = 0
         this.duration = 0
     }
@@ -43,7 +42,6 @@ class Socket {
             ioStream(socket).on("upload", (stream) => {
                 let filename = `${this.id}-${this.sequence}.ts`
                 let filePath = path.resolve(__dirname, `./uploads/${filename}`)
-                let tmpPath = `${filePath}_tmp`
                 /**
                  * 采用ffmpeg转webm格式为mpeg-ts片断
                  * 手动设置视频片段的offset
@@ -69,7 +67,7 @@ class Socket {
                             this.duration += data.format.duration
                             this.addVideo(`/uploads/${filename}`, data.format.duration)
                             this.sequence++
-                            if(this.sequence === MAX_VIDEO_FILES){
+                            if (this.sequence === MAX_VIDEO_FILES) {
                                 C.ready(this.id)
                                 // 刚开始直播的时候，由于服务器端还为处理完第一片视频，客户端是无法播放的。直邮等待服务器端处理好了视频之后才能播放
                                 socket.emit("videoReady")
@@ -81,21 +79,25 @@ class Socket {
             /**
              * 弹幕，接收客户端的message，然后直接广播出去
              */
-            socket.on('message', (data)=>{
+            socket.on('message', (data) => {
                 this.ns.emit('message', data)
             })
             /**
              * 当用户断开socket连接的时候，修改直播频道状态
              */
-            socket.on('disconnect', ()=>{
+            socket.on('disconnect', () => {
+                let channel = C.getOne(this.id)
+                if (!socket.user.id === channel.owner) {
+                    return
+                }
                 C.done(this.id)
                 /**
                  * 删除掉磁盘上的文件
                  */
-                this.videos.forEach((item)=>{
+                this.videos.forEach((item) => {
                     let file = path.resolve(__dirname, `.${item.video}`)
-                    fs.unlink(file, (err)=>{
-                        if(err){
+                    fs.unlink(file, (err) => {
+                        if (err) {
                             debug(`delete file: ${file} error`)
                         }
                     })
@@ -105,7 +107,7 @@ class Socket {
         })
     }
 
-    _remove(){
+    _remove() {
         let id = this.id
         sockets[id] = null
         delete sockets[id]
@@ -169,10 +171,13 @@ const getSocket = (id) => {
     return sockets[id]
 }
 
+const create = (id) => {
+    let s = new Socket(id)
+    sockets[id] = s
+}
+
 module.exports = {
     initIO,
     getSocket,
-    create(id) {
-        new Socket(id)
-    }
+    create
 }
